// lib/parser.gen.ts
export interface ASTNodeIntf {
    kind: ASTKinds;
}
export enum ASTKinds {
    start = "start",
    main_1 = "main_1",
    main_2 = "main_2",
    term_1 = "term_1",
    term_2 = "term_2",
    term_3 = "term_3",
    abstraction = "abstraction",
    application = "application",
    group = "group",
    identifier = "identifier",
    $EOF = "$EOF",
}
export interface start {
    kind: ASTKinds.start;
    main: main;
}
export type main = main_1 | main_2;
export type main_1 = application;
export type main_2 = term;
export type term = term_1 | term_2 | term_3;
export type term_1 = abstraction;
export interface term_2 {
    kind: ASTKinds.term_2;
    identifier: identifier;
}
export type term_3 = group;
export interface abstraction {
    kind: ASTKinds.abstraction;
    parameter: identifier;
    body: main;
}
export interface application {
    kind: ASTKinds.application;
    func: main;
    arg: term;
}
export interface group {
    kind: ASTKinds.group;
    group: main;
}
export type identifier = string;
// ...
export interface PosInfo {
    readonly overallPos: number;
    readonly line: number;
    readonly offset: number;
}
export interface RegexMatch {
    readonly kind: "RegexMatch";
    readonly negated: boolean;
    readonly literal: string;
}
export type EOFMatch = { kind: "EOF"; negated: boolean };
export type MatchAttempt = RegexMatch | EOFMatch;
export class SyntaxErr {
    public pos: PosInfo;
    public expmatches: MatchAttempt[];
    constructor(pos: PosInfo, expmatches: MatchAttempt[]) {
        this.pos = pos;
        this.expmatches = [...expmatches];
    }
    public toString(): string {
        return `Syntax Error at line ${this.pos.line}:${this.pos.offset}. Expected one of ${this.expmatches.map(x => x.kind === "EOF" ? " EOF" : ` ${x.negated ? 'not ': ''}'${x.literal}'`)}`;
    }
}